Split out FViewSteamId, fix imports
